reduce code smell

+++ WARDuino.h
@@ -187,7 +187,7 @@ class WARDuino {
 
    public:
     WARDuino();
-    int run_module(uint8_t *bytes, int size);
+    int run_module(uint8_t *bytes, size_t size);
     Module *load_module(uint8_t *bytes, uint32_t byte_count, Options options);
     bool invoke(Module *m, uint32_t fidx);
     uint32_t get_export_fidx(Module *m, const char *name);
